Improve reset password flow
